Updates knex migration file to check whether member paid table exists

+++ db/migrations/20231022015529_init.js
@@ -3,37 +3,41 @@
  * @returns { Promise<void> }
  */
 exports.up = function (knex) {
-  return knex.schema.createTable('member_paid', (table) => {
-    table.increments();
-    table.integer('chama_id').unsigned().notNullable();
-    table
-      .foreign('chama_id')
-      .references('id')
-      .inTable('chama')
-      .onDelete('CASCADE')
-      .onUpdate('CASCADE');
-    table.integer('member_id').unsigned().notNullable();
-    table
-      .foreign('member_id')
-      .references('id')
-      .inTable('members')
-      .onDelete('CASCADE')
-      .onUpdate('CASCADE');
-    table.integer('payment_transaction_id').unsigned().notNullable();
-    table
-      .foreign('payment_transaction_id')
-      .references('id')
-      .inTable('payment_transactions')
-      .onDelete('CASCADE')
-      .onUpdate('CASCADE');
-    table.integer('chama_cycle_count_id').unsigned().notNullable();
-    table
-      .foreign('chama_cycle_count_id')
-      .references('id')
-      .inTable('chama_cycle_count')
-      .onDelete('CASCADE')
-      .onUpdate('CASCADE');
-    table.timestamps(true, true);
+  return knex.schema.hasTable('member_paid').then(function (exists) {
+    if (!exists) {
+      return knex.schema.createTable('member_paid', (table) => {
+        table.increments();
+        table.integer('chama_id').unsigned().notNullable();
+        table
+          .foreign('chama_id')
+          .references('id')
+          .inTable('chama')
+          .onDelete('CASCADE')
+          .onUpdate('CASCADE');
+        table.integer('member_id').unsigned().notNullable();
+        table
+          .foreign('member_id')
+          .references('id')
+          .inTable('members')
+          .onDelete('CASCADE')
+          .onUpdate('CASCADE');
+        table.integer('payment_transaction_id').unsigned().notNullable();
+        table
+          .foreign('payment_transaction_id')
+          .references('id')
+          .inTable('payment_transactions')
+          .onDelete('CASCADE')
+          .onUpdate('CASCADE');
+        table.integer('chama_cycle_count_id').unsigned().notNullable();
+        table
+          .foreign('chama_cycle_count_id')
+          .references('id')
+          .inTable('chama_cycle_count')
+          .onDelete('CASCADE')
+          .onUpdate('CASCADE');
+        table.timestamps(true, true);
+      });
+    }
   });
 };
